Revert "nixos/release.nix: Pass a pkgs without aliases to tests"
We don't have a good way of adding `allowAliases = false` to an existing
`pkgs`, which means that we can't enable this on `nixosTests.*` because
that would break coherence between `pkgs` as in literally `pkgs.nixosTest`
and the `pkgs` module argument that is available there.

Having a discrepancy between the two means that neither OfBorg nor users
will catch the eval errors that makes tests effectively disappear from
Hydra. This is a Hydra UX issue, not a nixpkgs issue but a blocker
nonetheless.

This reverts commit 2d0a47236d5bda386de3f790e7a192cb42ea515c.